move streaming quality enum
initialize device profile properly in the device profiles store
