+++ src/enums/audio-quality.ts
@@ -1,0 +1,8 @@
+enum StreamingQuality {
+	Original = 'original', // Direct Play
+	High = 'high', // 320
+	Medium = 'medium', // 256
+	Low = 'low', // 128
+}
+
+export default StreamingQuality
